move part of checkoutDisplay into checkoutDetails to make things easier to deal with, create constants file for us state abbreviations, small style tweaks, setup potential models for user, customer address and order
