+++ prisma/schema.prisma
@@ -22,6 +22,8 @@ model Product {
   featured Boolean @default(false)
   category Category @relation(fields:[categoryId], references: [id]) 
   categoryId String
+  orders Order @relation(fields:[orderIds], references: [id])
+  orderIds String[]
   @@unique([name])
 }
 
@@ -33,3 +35,30 @@ model Category {
   @@unique([name])
 }
 
+model User {
+  id String @default(uuid()) @id
+  email String
+  addresses CustomerAddress[]
+  orders Order[]
+  @@unique([email])
+}
+
+model CustomerAddress {
+  id String @default(uuid()) @id
+  user User @relation(fields:[userId], references: [id])
+  userId String
+  firstName String
+  lastName String
+  AddressOne String
+  AddressTwo String?
+  City String
+  State String
+  ZipCode String
+}
+
+model Order {
+  id String @default(uuid()) @id
+  user User @relation(fields:[userId], references: [id])
+  userId String
+  products Product[]
+}
